Redid audio system
Added a SoundManager class that handles the playback and volume of all music and sound effects

From the outside the app appears the exact same as before.  But now instead of having to make a new mediaPlayer on each screen for the 'Click' sound we can instead just call SoundManager.playSfx(String name).  This also gives us one single point to control all music and sfx volumes